+++ changelog.md
@@ -3,3 +3,4 @@ Starting to keep track of changes as of October 28, 2015. I will try my best to 
 You can also browse the commit history to track my commits: [https://github.com/owenconti/livecodingtv-bot/commits/master](https://github.com/owenconti/livecodingtv-bot/commits/master)
 
 * Updated Loader to allow bot to start without a `/plugins` directory.
+* Verifying credentials.js includes the required attributes to start up when starting the bot.